refactor(registration-react): extract apiRegistrationConnector
